split 2nd group further, rename files
